# docs/swagger.yaml
definitions:
...
  apiAuth.paramsUserToken:
    properties:
      platform:
        type: integer
      secret:
        type: string
      uid:
        type: string
    required:
    - platform
    - secret
    - uid
    type: object
  apiAuth.resultUserRegister:
    properties:
      expiredTime:
        type: integer
      token:
        type: string
      uid:
        type: string
    type: object
...
  apiChat.paramsUserPullMsg:
    properties:
      data:
        properties:
          seqBegin:
            type: integer
          seqEnd:
            type: integer
        required:
        - seqBegin
        - seqEnd
        type: object
      operationID:
        type: string
      reqIdentifier:
        type: integer
      sendID:
        type: string
    required:
    - operationID
    - reqIdentifier
    - sendID
    type: object
  apiChat.paramsUserPullMsgBySeqList:
    properties:
      operationID:
        type: string
      reqIdentifier:
        type: integer
      sendID:
        type: string
      seqList:
        items:
          type: integer
        type: array
    required:
    - operationID
    - reqIdentifier
    - sendID
    type: object
  apiChat.paramsUserSendMsg:
    properties:
      data:
        properties:
          clientMsgID:
            type: string
          content:
            type: string
          contentType:
            type: integer
          ext:
            additionalProperties: true
            type: object
          forceList:
            items:
              type: string
            type: array
          msgFrom:
            type: integer
          offlineInfo:
            additionalProperties: true
            type: object
          options:
            additionalProperties: true
            type: object
          recvID:
            type: string
          sessionType:
            type: integer
        required:
        - clientMsgID
        - content
        - contentType
        - msgFrom
        - recvID
        - sessionType
        type: object
      operationID:
        type: string
      platformID:
        type: integer
      reqIdentifier:
        type: integer
      sendID:
        type: string
      senderFaceUrl:
        type: string
      senderNickName:
        type: string
    required:
    - operationID
    - platformID
    - reqIdentifier
    - sendID
    type: object
  apiChat.resultSendMsg:
    properties:
      clientMsgID:
        type: string
      sendTime:
        type: integer
      serverMsgID:
        type: string
    type: object
...
  friend.paramsSearchFriend:
    properties:
      operationID:
        type: string
      ownerUid:
        type: string
      uid:
        type: string
    required:
    - operationID
    - uid
    type: object
...
  user.result:
    properties:
      data: {}
      errCode:
        example: 0
        type: integer
      errMsg:
        example: error
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /auth/user_token:
    post:
      consumes:
      - application/json
      description: get token
      parameters:
      - description: get token params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apiAuth.paramsUserToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                data:
                  $ref: '#/definitions/apiAuth.resultUserRegister'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - auth
...
  /chat/pull_msg:
    post:
      consumes:
      - application/json
      description: user pull messages
      parameters:
      - description: user pull messages
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apiChat.paramsUserPullMsg'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                reqIdentifier:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - chat
  /chat/pull_msg_by_seq:
    post:
      consumes:
      - application/json
      description: user pull msg by seq
      parameters:
      - description: pull msg by seq
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apiChat.paramsUserPullMsgBySeqList'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                reqIdentifier:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - chat
  /chat/send_msg:
    post:
      consumes:
      - application/json
      description: user send messages
      parameters:
      - description: user send messages
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apiChat.paramsUserSendMsg'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                data:
                  $ref: '#/definitions/apiChat.resultSendMsg'
                reqIdentifier:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - chat
  /friend/add_blacklist:
    post:
      consumes:
      - application/json
      description: add a user into black list
      parameters:
      - description: add black list params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsSearchFriend'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
...
swagger: "2.0"